@@ -1,6 +1,6 @@
-require_relative 'boot'
+require_relative "boot"
 
-require 'rails/all'
+require "rails/all"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -9,16 +9,25 @@ Bundler.require(*Rails.groups)
 module BioportalWebUi
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
+    config.load_defaults 6.1
+    config.active_record.yaml_column_permitted_classes = [
+      ActionController::Parameters,
+      HashWithIndifferentAccess
+    ]
 
-    # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration can go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded after loading
-    # the framework and any gems in your application.
+    # Configuration for the application, engines, and railties goes here.
+    #
+    # These settings can be overridden in specific environments using the files
+    # in config/environments, which are processed later.
+    #
+    # config.time_zone = "Central Time (US & Canada)"
+    # config.eager_load_paths << Rails.root.join("extras")
 
-    # Serve error pages from the Rails app itself, instead of using static error pages in /public.
     config.exceptions_app = self.routes
 
-    config.settings = config_for(:settings)
+    config.settings = config_for :settings
+
+    # Initialize configuration for KGCL change request functionality.
+    config.change_request = config_for :change_request
   end
 end
